feat: make team reviews a parameter (#14)
Previously, all PRs automatically added the sourcegrapm/team/release
reviewer to PRs. That worked fine while we were only using this on
Monorepo, but now we're expanding usage to Cody, Jetbrains and the
Deploy-* repos. Cody and Jetbrains particularly do not have the release
team as collaborator which caused errors while using this automation.

We're keeping the default the same with this change, but now we can
optionally override the default to change the team, or remove it
entirely (setting teamReviews = "").